Added options for dual nozzle conversion
Added options to the modified heatsink duct and the insulator to suit
the dual nozzle conversion. The standoffs on the heatsink duct for
fitting the sensor board need to be deeper for the dual nozzle, to give
adequate clearance between the sensor components and the nearer of the
two heater blocks. The insulator pillars are made thinner to compensate
in part.