Updated firmware for dual board
Updated the firmware for the dual nozzle board to use the same offset
cancellation technique as the single nozzle board when reading the
thermistors. Reduced the number of thermistor samples averaged for the
dual board to avoid running out of RAM. Also added holders for dual
nozzle and mini boards.